Rename winuserspace to netifuserspace
