+++ pkg/proxy/netifuserspace/BUILD
@@ -15,7 +15,7 @@ go_library(
         "roundrobin.go",
         "types.go",
     ],
-    importpath = "k8s.io/kubernetes/pkg/proxy/winuserspace",
+    importpath = "k8s.io/kubernetes/pkg/proxy/netifuserspace",
     deps = [
         "//pkg/apis/core/v1/helper:go_default_library",
         "//pkg/proxy:go_default_library",
