+++ pkg/proxy/netifuserspace/loadbalancer.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package winuserspace
+package netifuserspace
 
 import (
 	"k8s.io/api/core/v1"
